+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+- Added: `wp s1 generate block:middleware <name>` CLI command to create sample block middleware.
+- Added: `wp s1 generate block <name> --with-middleware` that adds the appropriate interface and a stub middleware parameters method to a generated block config.
+- Updated: Block model generator template `model.php.tmpl` will use a different method name depending on if block middleware is detected in the version of Square One this library is included in.
+- Added: `With_Field_Finder` trait, that can search a block object's fields for a specific field key.
 
 ## 3.4.18 - 2022-06-30
 - Updated: `so wp s1 generate block <name>` will use `esc_html__()` instead of `__()` for field labels.